Fix Linter Erros & Add schema.sql File
