+++ .rubocop.yml
@@ -3,8 +3,6 @@ AllCops:
   Exclude:
     - "Guardfile"
     - "Rakefile"
-    - "main.rb"
-    - "app.rb"
     - "node_modules/**/*"
 
   DisplayCopNames: true
@@ -12,7 +10,7 @@ AllCops:
 Layout/LineLength:
   Max: 120
 Metrics/MethodLength:
-  Max: 20
+  Max: 30
 Metrics/AbcSize:
   Max: 50
 Metrics/ClassLength:
@@ -50,4 +48,6 @@ Style/HashEachMethods:
 Style/HashTransformKeys:
   Enabled: false
 Style/HashTransformValues:
+  Enabled: false
+Lint/ToJSON:
   Enabled: false
