+++ Book/book.json
@@ -1,0 +1,1 @@
+[{"id":868,"publisher":"publisher Printer","cover_state":"good","publish_date":"2018-02-02"},{"id":21,"publisher":"new Publisher","cover_state":"bad","publish_date":"2002-02-01"}]
