+++ app.rb
@@ -1,5 +1,5 @@
 require 'fileutils'
-require_relative 'book_ui'
+require_relative './Book/book_ui'
 require_relative './musicAlbum/music_album_ui'
 require 'JSON'
