--- book.json
@@ -1,1 +1,0 @@
-[{"id":767,"publisher":"Syed Pritner","cover_state":"bad","publish_date":"2020-02-25"},{"id":637,"publisher":"Atril","cover_state":"good","publish_date":"2002-02-02"},{"id":939,"publisher":"wajad","cover_state":"good","publish_date":"2008-08-15"},{"id":375,"publisher":"syed","cover_state":"bad","publish_date":"2003-03-31"},{"id":285,"publisher":"syed","cover_state":"good","publish_date":"2003-03-31"},{"id":25,"publisher":"wajad","cover_state":"good","publish_date":"2008-08-01"}]
