--- label.json
@@ -1,1 +1,0 @@
-[{"id":987,"title":null,"color":null},{"id":247,"title":"Boy","color":"White"}]
