+++ musicAlbum/music_album.json
@@ -1,1 +1,1 @@
-[{"publish_date":"202-12-12","on_spotify":true,"genre":"song"},{"publish_date":"2022-02-22","on_spotify":true,"genre":"syed"}]
+[{"publish_date":"2022-02-02","on_spotify":false,"genre":"syed"}]
